Fixed behaviour with pandoc 1.9.3 and 1.9.1.2
As suggested in the commit message, although
eee909faacace8f509fc960adfad02c0d600a952 fixed the code for pandoc
1.10.1 and newer, it didn't then work for 1.9.3 and older.

But it gave me enough of a clue to be able to adjust the code to now
work with all versions.

I've updated the README.md to now list the 4 pandoc versions it works
fine on now.

So this fully fixes issue #1